Clang apparently doesn't respond to -falign-loops or -finline-limit. Removing them from OS X project files. Quiet some warnings about implicit conversion from 64-bit integer to 32-bit in the main app, but only the Cocoa parts.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@85897 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/ConfigurationController.m
@@ -83,7 +83,7 @@ NSString *const memCardChangeNumberKey = @"PcsxrMemoryCardThatChangedKey";
 	if ([openDlg runModal] == NSFileHandlingPanelOKButton) {
 		NSURL *mcdURL = [[openDlg URLs] objectAtIndex:0];
         
-		[ConfigurationController setMemoryCard:tag toURL:mcdURL];
+		[ConfigurationController setMemoryCard:(int)tag toURL:mcdURL];
     }
 	RELEASEOBJ(openDlg);	
 }
@@ -114,14 +114,14 @@ NSString *const memCardChangeNumberKey = @"PcsxrMemoryCardThatChangedKey";
 		strlcpy(mcd, [[mcdURL path] fileSystemRepresentation], MAXPATHLEN);
 		CreateMcd(mcd);
 		
-		[ConfigurationController setMemoryCard:tag toURL:mcdURL];
+		[ConfigurationController setMemoryCard:(int)tag toURL:mcdURL];
     }
 	RELEASEOBJ(openDlg);
 }
 
 - (IBAction)setVideoType:(id)sender
 {
-	int tag = [[sender selectedItem] tag];
+	NSInteger tag = [[sender selectedItem] tag];
 
 	if (3 == tag) {
 		[[NSUserDefaults standardUserDefaults] setBool:YES forKey:@"AutoDetectVideoType"];
@@ -193,7 +193,7 @@ NSString *const memCardChangeNumberKey = @"PcsxrMemoryCardThatChangedKey";
 
 	// setup labels
 
-	int tag = [defaults integerForKey:@"AutoDetectVideoType"];
+	NSInteger tag = [defaults integerForKey:@"AutoDetectVideoType"];
 	if (tag)
 		tag = 3;
 	else {
